fixed a problem at index found another problem while getting phase in createFromVector

+++ TeamViewer/packetParser.cpp
@@ -120,7 +120,7 @@ std::unique_ptr<const DefaultControlPacket> PacketParser::createDefaultControlPa
     ControlPacketTypes controlPacketType;
     std::unique_ptr<const DefaultPacket> defaultPacket = createDefaultPacketFromVector(defaultControlPacketBuffer, index);
     controlPacketType = static_cast<ControlPacketTypes>(networkToHost<uint16_t>(defaultControlPacketBuffer, index));
-    index += sizeof(uint32_t);
+    index += sizeof(uint16_t);
     std::unique_ptr<const DefaultControlPacket> packet = std::make_unique<DefaultControlPacket>(defaultPacket->getAckSequenceNumber(), defaultPacket->getPacketSequenceNumber(), defaultPacket->getTimeStamp(), controlPacketType);
     return packet;
 }
